+++ Assets/_Scripts/Characters/NPC/Base Scripts/Interfaces/IDamageble.cs
@@ -2,19 +2,16 @@ using UnityEngine;
 
 public interface IDamageble
 {
-    
-
-    int MaxHealth { get; set; }
-    int CurrentHealth { get; set; }
+    float MaxHealth { get; set; }
+    float CurrentHealth { get; set; }
     float RegenDelay { get; set; }
-    int RegenRate { get; set; }
+    float RegenRate { get; set; }
 
     bool Alive { get; set; }
     bool TakingDamage { get; set; }
     int GotDamagedCounter { get; set; }
 
-    void TakeDamage(int amount);
+    void TakeDamage(float amount);
     void Die();
     void SetInitialIDamageble(SafeInstantiation.HealthStats? healthStats);
-
 }
